+++ redwood-tooling-lint/build.gradle
@@ -1,3 +1,6 @@
+import org.jetbrains.kotlin.gradle.dsl.JvmTarget
+import org.jetbrains.kotlin.gradle.tasks.KotlinJvmCompile
+
 apply plugin: 'org.jetbrains.kotlin.jvm'
 apply plugin: 'org.jetbrains.kotlin.plugin.serialization'
 apply plugin: 'com.vanniktech.maven.publish'
@@ -12,11 +15,23 @@ dependencies {
   implementation libs.clikt
   implementation libs.kotlinx.serialization.core
   implementation libs.xmlutil.serialization
+  implementation libs.lint.core
 
   testImplementation libs.kotlin.test
   testImplementation libs.junit
   testImplementation libs.assertk
   testImplementation libs.jimfs
+}
+
+tasks.withType(JavaCompile).configureEach {
+  sourceCompatibility = JavaVersion.VERSION_11.toString()
+  targetCompatibility = JavaVersion.VERSION_11.toString()
+}
+
+tasks.withType(KotlinJvmCompile).configureEach {
+  compilerOptions {
+    jvmTarget.set(JvmTarget.JVM_11)
+  }
 }
 
 tasks.named("distTar").configure { task ->
